+++ client/ntlogin.go
@@ -13,8 +13,6 @@ import (
 	"github.com/LagrangeDev/LagrangeGo/utils/crypto"
 )
 
-var loginLogger = utils.GetLogger("login")
-
 func buildNtloginCaptchaSubmit(ticket, randStr, aid string) proto.DynamicMessage {
 	return proto.DynamicMessage{
 		1: ticket,
@@ -50,7 +48,6 @@ func buildNtloginRequest(uin uint32, app *auth.AppInfo, device *auth.DeviceInfo,
 		body[1].(proto.DynamicMessage)[5] = proto.DynamicMessage{1: sig.Cookies}
 	}
 	if all(sig.CaptchaInfo[:3]) {
-		loginLogger.Debugln("login with captcha info")
 		body[2].(proto.DynamicMessage)[2] = buildNtloginCaptchaSubmit(sig.CaptchaInfo[0], sig.CaptchaInfo[1], sig.CaptchaInfo[2])
 	}
 
@@ -66,7 +63,7 @@ func buildNtloginRequest(uin uint32, app *auth.AppInfo, device *auth.DeviceInfo,
 	}.Encode(), nil
 }
 
-func ParseNtloginResponse(response []byte, sig *auth.SigInfo) (loginState.State, error) {
+func parseNtloginResponse(response []byte, sig *auth.SigInfo) (loginState.State, error) {
 	var frame login.SsoNTLoginEncryptedData
 	err := proto.Unmarshal(response, &frame)
 	if err != nil {
@@ -93,28 +90,23 @@ func ParseNtloginResponse(response []byte, sig *auth.SigInfo) (loginState.State,
 		sig.D2 = body.Credentials.D2
 		sig.D2Key = body.Credentials.D2Key
 		sig.TempPwd = body.Credentials.TempPassword
-		loginLogger.Debugln("SigInfo got")
 		return loginState.Success, nil
-	} else {
-		ret := loginState.State(base.Header.Error.ErrorCode)
-		if ret == loginState.CaptchaVerify {
-			sig.Cookies = base.Header.Cookie.Cookie.Unwrap()
-			verifyUrl := body.Captcha.Url
-			aid := getAid(verifyUrl)
-			sig.CaptchaInfo[2] = aid
-			loginLogger.Warningln("need captcha verify: ", verifyUrl)
-		} else if base.Header.Error.Tag != "" {
-			stat := base.Header.Error
-			title := stat.Tag
-			content := stat.Message
-			loginLogger.Errorf("Login fail on ntlogin(%s): [%s]>%s", ret.Name(), title, content)
-			return -999, fmt.Errorf("login fail on ntlogin(%s): [%s]>%s", ret.Name(), title, content)
-		} else {
-			loginLogger.Errorf("Login fail: %s", ret.Name())
-			return -999, fmt.Errorf("login fail: %s", ret.Name())
-		}
 	}
-	return loginState.State(base.Header.Error.ErrorCode), nil
+	ret := loginState.State(base.Header.Error.ErrorCode)
+	if ret == loginState.CaptchaVerify {
+		sig.Cookies = base.Header.Cookie.Cookie.Unwrap()
+		verifyUrl := body.Captcha.Url
+		aid := getAid(verifyUrl)
+		sig.CaptchaInfo[2] = aid
+		return ret, fmt.Errorf("need captcha verify: %v", verifyUrl)
+	}
+	if base.Header.Error.Tag != "" {
+		stat := base.Header.Error
+		title := stat.Tag
+		content := stat.Message
+		return ret, fmt.Errorf("login fail on ntlogin(%s): [%s]>%s", ret.Name(), title, content)
+	}
+	return ret, fmt.Errorf("login fail: %s", ret.Name())
 }
 
 func getAid(verifyUrl string) string {
